functional, untested

+++ connect_scripts/add-extensions-all.py
@@ -3,6 +3,8 @@ import json
 from boto3.dynamodb.conditions import Key,Attr
 
 # TODO: function to remove user, exception handling, event handler for lambda
+# , backoff algorithm for connect api request in get_users, remove users
+# not used
  
 dynamodb = boto3.client('dynamodb', region_name = 'us-east-1') 
 res_dynamodb = boto3.resource('dynamodb')
@@ -14,23 +16,21 @@ unused = []
 
 
 # checks if a user is currently assigned an extension
-#TODO: test 
 def has_extension(username):
     u = username
 
     response = table.query(
-            IndexName='byAgent',
+            IndexName='skIndex',
             KeyConditionExpression=Key('sk').eq('agentID') & Key('sk_value').eq(u))
     hasExt = response.get('Items', {})
-
+# checks if the stored query result (hasExt) does not contain an item, in which case it returns false
     if (bool(hasExt)):
-        return False 
-    else:
         return True 
+    else:
+        return False 
         
 
 # gets list of connect users, returns as list of dicts
-# TODO: test
 def  get_users():
     users =[] 
     iid ='794790f5-0ae2-4348-806a-f58bf245ab3f'
@@ -42,23 +42,26 @@ def  get_users():
     while "NextToken" in response:
         addUsers = response.get('UserSummaryList')
 
-        users.append(addUsers)
+        for _ in addUsers:
+            users.append(_.get('Username'))
 
         response = connect.list_users(
                 NextToken=response["NextToken"],
                 InstanceId=iid,
                 MaxResults=10)
 
-    users.append(response.get('UserSummaryList'))
-    
+    for _ in response.get('UserSummaryList'):
+        users.append(_.get('Username'))
     breakpoint()
+    
     return users
 
 
 # gets "not used (nu)" extensions 100 at a time
-#TODO: add while loop for when unused is <100 to repeat query. Only return pk and sk as sk_value is the same as pk for these items
+# TODO: raise exception when unable to find nu items
 def get_unused_ext():
     global unused
+
     response = table.query(
     IndexName='skIndex',
     Limit=100,
@@ -71,16 +74,18 @@ def get_unused_ext():
         
     return 
 
-
+# TODO: return exception if user is already assigned an extension
 def set_extension(username):
     u = username
     global unused
 
+    if has_extension(username):
+        return
     if not bool(unused):
         get_unused_ext()
 
     extension = unused.pop()
-# TODO: make transactional delete checking for any other items using same 'pk'
+# TODO: make conditional delete to verify that 'nu' item still exists
     response = table.delete_item(
         Key={
             'pk': extension,
@@ -96,9 +101,16 @@ def set_extension(username):
             }
     )
 
-    breakpoint()
     return
 
-#get_unused_ext()
-set_extension("mmendez")
-breakpoint()
+def update_db():
+    users = get_users() 
+    for _ in users
+        set_extension(_)
+    return
+
+
+def lambda_handler(event, context):
+    response = update_db()
+    return response
+
